requirements.txt
aiohttp==3.6.2
alive-progress==1.6.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.2.7
et-xmlfile==1.0.1
geoip2==4.0.2
gunicorn==20.0.4
idna==2.9
whitenoise==5.0.1
django-cors-headers==3.10.0
django-environ==0.8.1
django-import-export==2.0.2
django-map-widgets==0.3.0
django-mathfilters==1.0.0
django-rest-passwordreset==1.2.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
python-decouple==3.5
